requirements.txt
Flask==3.1.3
paho-mqtt==1.6.1
